ISSUE #268 * Improve tests structure.
